cv_mit_threads

+++ src/utils/Datapoint.java
@@ -10,22 +10,22 @@ public class Datapoint {
     private int label;
     
     //values for the datapoint
-    private double[] values;
+    private int[] values;
  
-    public Datapoint(int label, double[] values) {
+    public Datapoint(int label, int[] values) {
         this.label = label;
         this.values = values;
     }
     
     public Datapoint(int sizeOfData) {
-        values = new double[sizeOfData];
+        values = new int[sizeOfData];
     }
     
-    public void addValue(int idx, double value){values[idx] = value;}
+    public void addValue(int idx, int value){values[idx] = value;}
     public void setLabel(int label) {this.label = label;}
-    public void setValues(double[] values) {this.values = values;}
+    public void setValues(int[] values) {this.values = values;}
 
     public int getLabel(){return label;}
-    public double[] getValues(){return values;}
+    public int[] getValues(){return values;}
     
 }
